fix: respect and prefill provided nickname from registration class
Today if you supply through an OAuth2 client a nickname, the signup modal will not respect this and still override the nickname input field with the username. 

However with this fix, you can get OAuth2 to put `username` based on id and `nickname ` with a UTF-8 characters and no restrictions.